inventory_service/inventory/serializers.py
from rest_framework import serializers

from .models import Product, Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'


class ProductSerializer(serializers.ModelSerializer):
    category = CategorySerializer()

    class Meta:
        model = Product
        fields = '__all__'


class ProductsByCategoryQuerySerializer(serializers.Serializer):
    category_name = serializers.CharField(required=False, allow_null=True)
